requirements.txt
aiohttp==3.4.4
aiohttp-auth==0.1.1
aiohttp-route-decorator==0.1.4
aiohttp-security==0.4.0
aiohttp-session==2.7.0
aiohttp-swagger==1.0.5
aiojobs==0.2.2
aiosqlite==0.7.0
cryptography==2.3.1
voluptuous==0.11.5
pyfigle==0.7.6